Revert all changes related to storing and restoring non-rel data in page server
This includes the following commits:

1a3a572c6eaba3f43f43200d049a1e451c70f432 Specify right LSN in test_createdb.py
629c4c177981b63651af145e4e74dfe9398c032f Fix issue with propagation of CREATE DATABASE to the branch
5c976b4deeb843a61a5699d4777d0bc49cc08ea0 [refer #167] Fix handling of pg_filenode.map files in page server
c1edb79e1e117b5e22b1b77197d0286937036646 Fix merge conflict: set initial WAL position to second segment because of pg_resetwal
ae7a633cc7145fab3211771d577b8e6d027532fb Add nonrelfile_utils.rs file
8d7f7156462e56b00dd51f3faa6f57adbc634988 Fix restoring non-relational data during compute node startup
834db25c6c32cca6e3f2df1386f767825c773b4a Do not transfer WAL to computation nodes: use pg_resetwal for node startup

As well as some older changes related to storing CLOG and MultiXact data as
"pseudorelation" in the page server.

With this revert, we go back to the situtation that when you create a
new compute node, we ship *all* the WAL from the beginning of time to
the compute node. Obviously we need a better solution, like the code
that this reverts. But per discussion with Konstantin and Stas, this
stuff was still half-baked, and it's better for it to live in a branch
for now, until it's more complete and has gone through some review.