Try to silence test_gc_cutoff flakiness.
Commit 6308a4598d changed the test, so that it fails if pgbench runs
to completion without triggering the failpoint. That has now happened
several times in the CI. That's not expected, so this needs some
investigation, but as a quick fix just make the pgbench runs longer so
that we're closer to the situation before commit 6308a4598d.

See https://github.com/neondatabase/neon/issues/2856