test: adjust checkpoint distance in `test_layer_map` (#6842)
https://github.com/neondatabase/neon/commit/ff73f2c7dc7e42f57a6763057126dc70c4f4e036
changed the layer rolling logic to more closely obey the
`checkpoint_distance` config. Previously, this test was getting
layers significantly larger than the 8K it was asking for. Now the
payload in the layers is closer to 8K (which means more layers in
total).

Tweak the `checkpoint_distance` to get a number of layers more
reasonable for this test. Note that we still get more layers than
before (~8K vs ~5K).